+++ src/main/java/com/example/moneytrack/dto/AccountCreateRequest.java
@@ -3,12 +3,10 @@ package com.example.moneytrack.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Getter
 @NoArgsConstructor
 public class AccountCreateRequest {
     private String productCode;
     private Integer memberId;
-    private BigDecimal deposit;
+    private Long deposit;
 }
